Changed similarity to be calculated systematically
As the bounding boxes are already found in descending order of min Y values. It can be assumed that each box in the images can be paired to its equivalent box in the model by its index.

Also changed the similarity to instead of being the average over all the boxes to intead be totalOverlapArea/totalBoxesArea