Added StateWeightedLevenshtein alg for word similarity
I found using a weighted levenshtein edit distance but instead of computing it with characters but instead the state sequence achieves a good metric for spell correction confidence levels. This is because it accomadates for the way our HMM treats insertions and deletions as 'ol', 'al' etc is seen as 1 state but 2 characters. So normal WL alg wouldn't accomodate for this.