Corrected HMM and Change UMLS search
HMM wasn't valid as when a char was not encountered the whole column was 0.0. So now we set one of the values too 1.0 and the rest 0.0 in that column to make it valid.

Also changed the UMLS search to find the 3 closests matches to the word. Then we keep the result thats name is the closests SWL similarity to the orignal word.